service-rest/src/test/java/com/epam/brest/service/rest/DepartmentServiceRestTest.java
package com.epam.brest.service.rest;

import com.epam.brest.model.Department;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.apache.commons.lang3.RandomStringUtils;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpMethod;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.client.ExpectedCount;
import org.springframework.test.web.client.MockRestServiceServer;
import org.springframework.web.client.RestTemplate;

import java.net.URI;
import java.util.Arrays;
import java.util.List;

import static com.epam.brest.model.constants.DepartmentConstants.DEPARTMENT_NAME_SIZE;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.springframework.test.web.client.match.MockRestRequestMatchers.method;
import static org.springframework.test.web.client.match.MockRestRequestMatchers.requestTo;
import static org.springframework.test.web.client.response.MockRestResponseCreators.withStatus;

@ExtendWith(SpringExtension.class)
@ContextConfiguration(locations = {"classpath:app-context-test.xml"})
public class DepartmentServiceRestTest {

    private static final Logger LOGGER = LoggerFactory.getLogger(DepartmentServiceRestTest.class);

    public static final String DEPARTMENTS_URL = "http://localhost:8088/departments";

    @Autowired
    RestTemplate restTemplate;

    private MockRestServiceServer mockServer;

    private ObjectMapper mapper = new ObjectMapper();

    DepartmentServiceRest departmentService;

    @BeforeEach
    public void before() {
        mockServer = MockRestServiceServer.createServer(restTemplate);
        departmentService = new DepartmentServiceRest(DEPARTMENTS_URL, restTemplate);
    }

    @Test
    public void shouldFindAllDepartments() throws Exception {

        LOGGER.debug("shouldFindAllDepartments()");
        // given
        mockServer.expect(ExpectedCount.once(), requestTo(new URI(DEPARTMENTS_URL)))
                .andExpect(method(HttpMethod.GET))
                .andRespond(withStatus(HttpStatus.OK)
                        .contentType(MediaType.APPLICATION_JSON)
                        .body(mapper.writeValueAsString(Arrays.asList(create(0), create(1))))
                );

        // when
        List<Department> departments = departmentService.findAll();

        // then
        mockServer.verify();
        assertNotNull(departments);
        assertTrue(departments.size() > 0);
    }

    @Test
    public void shouldCreateDepartment() throws Exception {

        LOGGER.debug("shouldCreateDepartment()");
        // given
        Department department = new Department()
                .setDepartmentName(RandomStringUtils.randomAlphabetic(DEPARTMENT_NAME_SIZE));

        mockServer.expect(ExpectedCount.once(), requestTo(new URI(DEPARTMENTS_URL)))
                .andExpect(method(HttpMethod.POST))
                .andRespond(withStatus(HttpStatus.OK)
                        .contentType(MediaType.APPLICATION_JSON)
                        .body(mapper.writeValueAsString("1"))
                );
        // when
        Integer id = departmentService.create(department);

        // then
        mockServer.verify();
        assertNotNull(id);
    }

    @Test
    public void shouldFindDepartmentById() throws Exception {

        // given
        Integer id = 1;
        Department department = new Department()
                .setDepartmentId(id)
                .setDepartmentName(RandomStringUtils.randomAlphabetic(DEPARTMENT_NAME_SIZE));

        mockServer.expect(ExpectedCount.once(), requestTo(new URI(DEPARTMENTS_URL + "/" + id)))
                .andExpect(method(HttpMethod.GET))
                .andRespond(withStatus(HttpStatus.OK)
                        .contentType(MediaType.APPLICATION_JSON)
                        .body(mapper.writeValueAsString(department))
                );

        // when
        Department resultDepartment = departmentService.getDepartmentById(id);

        // then
        mockServer.verify();
        assertNotNull(resultDepartment);
        assertEquals(resultDepartment.getDepartmentId(), id);
        assertEquals(resultDepartment.getDepartmentName(), department.getDepartmentName());
    }

    @Test
    public void shouldUpdateDepartment() throws Exception {

        // given
        Integer id = 1;
        Department department = new Department()
                .setDepartmentId(id)
                .setDepartmentName(RandomStringUtils.randomAlphabetic(DEPARTMENT_NAME_SIZE));

        mockServer.expect(ExpectedCount.once(), requestTo(new URI(DEPARTMENTS_URL)))
                .andExpect(method(HttpMethod.PUT))
                .andRespond(withStatus(HttpStatus.OK)
                        .contentType(MediaType.APPLICATION_JSON)
                        .body(mapper.writeValueAsString("1"))
                );

        mockServer.expect(ExpectedCount.once(), requestTo(new URI(DEPARTMENTS_URL + "/" + id)))
                .andExpect(method(HttpMethod.GET))
                .andRespond(withStatus(HttpStatus.OK)
                        .contentType(MediaType.APPLICATION_JSON)
                        .body(mapper.writeValueAsString(department))
                );

        // when
        int result = departmentService.update(department);
        Department updatedDepartment = departmentService.getDepartmentById(id);

        // then
        mockServer.verify();
        assertTrue(1 == result);

        assertNotNull(updatedDepartment);
        assertEquals(updatedDepartment.getDepartmentId(), id);
        assertEquals(updatedDepartment.getDepartmentName(), department.getDepartmentName());
    }

    @Test
    public void shouldDeleteDepartment() throws Exception {

        // given
        Integer id = 1;
        mockServer.expect(ExpectedCount.once(), requestTo(new URI(DEPARTMENTS_URL + "/" + id)))
                .andExpect(method(HttpMethod.DELETE))
                .andRespond(withStatus(HttpStatus.OK)
                        .contentType(MediaType.APPLICATION_JSON)
                        .body(mapper.writeValueAsString("1"))
                );
        // when
        int result = departmentService.delete(id);

        // then
        mockServer.verify();
        assertTrue(1 == result);
    }

    private Department create(int index) {
        Department department = new Department();
        department.setDepartmentId(index);
        department.setDepartmentName("d" + index);
        return department;
    }
}
